moved code to file PackageDownloadResponse.swift
